requirements.txt
pyqt6<=6.8.0
pyserial<=3.5
